+++ Library/src/main/java/dev/testify/internal/extensions/InstrumentationRegistryExtensions.kt
@@ -26,6 +26,7 @@ package dev.testify.internal.extensions
 import android.app.Instrumentation
 import android.os.Bundle
 import androidx.test.platform.app.InstrumentationRegistry
+import dev.testify.annotation.ScreenshotInstrumentation
 
 class TestInstrumentationRegistry {
     companion object {
@@ -69,6 +70,18 @@ class TestInstrumentationRegistry {
         }
     }
 }
+
+/**
+ * Check if Testify is being run from the Gradle plugin.
+ */
+fun isInvokedFromPlugin(): Boolean =
+    InstrumentationRegistry.getArguments().containsKey("annotation")
+
+/**
+ * Returns the fully qualified dot-separated name of the annotation required by the Gradle plugin.
+ */
+fun getScreenshotAnnotationName(): String =
+    InstrumentationRegistry.getArguments().getString("annotation", ScreenshotInstrumentation::class.qualifiedName)
 
 private const val ESC_YELLOW = "${27.toChar()}[33m"
 private const val ESC_CYAN = "${27.toChar()}[36m"